# webapp/frontend/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HpoDetail": {
        "properties": {
          "best_config": {
            "additionalProperties": true,
            "title": "Best Config",
            "type": "object"
          },
          "best_config_reserves_holdout": {
            "title": "Best Config Reserves Holdout",
            "type": "boolean"
          },
          "best_trial_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Trial Number"
          },
          "best_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Value"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "direction": {
            "$ref": "#/components/schemas/StudyDirection"
          },
          "launched_by_username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Launched By Username"
          },
          "live_job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Live Job Id"
          },
          "n_complete": {
            "title": "N Complete",
            "type": "integer"
          },
          "n_trials": {
            "title": "N Trials",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "store": {
            "title": "Store",
            "type": "string"
          },
          "wire_id": {
            "title": "Wire Id",
            "type": "string"
          }
        },
        "required": [
          "wire_id",
          "name",
          "store",
          "created_at",
          "n_trials",
          "n_complete",
          "best_value",
          "best_trial_number",
          "direction",
          "best_config",
          "best_config_reserves_holdout",
          "live_job_id"
        ],
        "title": "HpoDetail",
        "type": "object"
      },
      "HpoSummary": {
        "properties": {
          "best_config_reserves_holdout": {
            "title": "Best Config Reserves Holdout",
            "type": "boolean"
          },
          "best_trial_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Trial Number"
          },
          "best_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Value"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "direction": {
            "$ref": "#/components/schemas/StudyDirection"
          },
          "has_best_config": {
            "title": "Has Best Config",
            "type": "boolean"
          },
          "launched_by_username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Launched By Username"
          },
          "n_complete": {
            "title": "N Complete",
            "type": "integer"
          },
          "n_trials": {
            "title": "N Trials",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "store": {
            "title": "Store",
            "type": "string"
          },
          "wire_id": {
            "title": "Wire Id",
            "type": "string"
          }
        },
        "required": [
          "wire_id",
          "name",
          "store",
          "created_at",
          "n_trials",
          "n_complete",
          "best_value",
          "best_trial_number",
          "direction",
          "has_best_config",
          "best_config_reserves_holdout"
        ],
        "title": "HpoSummary",
        "type": "object"
      },
...
      "ParamImportanceResponse": {
        "description": "Per-hyperparameter relative importance for the live HPO monitor.\n\n``importance`` is empty and ``message`` is set when the study has too few\ncompleted trials, the optuna DB is missing, or the importance evaluator\nraises on a degenerate search space - the endpoint stays 200 so the\nfrontend renders an empty-state card instead of erroring.",
        "properties": {
          "importance": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Importance",
            "type": "object"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "required": [
          "importance"
        ],
        "title": "ParamImportanceResponse",
        "type": "object"
      },
...
      "StudyConsolidatedDTO": {
        "description": "Read view of a study's consolidated report (tables + plots tree).",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "git_sha": {
            "title": "Git Sha",
            "type": "string"
          },
          "incomplete_leg_ids": {
            "items": {
              "type": "string"
            },
            "title": "Incomplete Leg Ids",
            "type": "array"
          },
          "n_legs_with_holdout": {
            "title": "N Legs With Holdout",
            "type": "integer"
          },
          "n_universes_with_pairwise": {
            "title": "N Universes With Pairwise",
            "type": "integer"
          },
          "plots": {
            "items": {
              "type": "string"
            },
            "title": "Plots",
            "type": "array"
          },
          "publish_label": {
            "title": "Publish Label",
            "type": "string"
          },
          "strategies": {
            "items": {
              "type": "string"
            },
            "title": "Strategies",
            "type": "array"
          },
          "study_name": {
            "title": "Study Name",
            "type": "string"
          },
          "tables": {
            "items": {
              "type": "string"
            },
            "title": "Tables",
            "type": "array"
          },
          "universes": {
            "items": {
              "type": "string"
            },
            "title": "Universes",
            "type": "array"
          }
        },
        "required": [
          "study_name",
          "publish_label",
          "created_at",
          "git_sha",
          "strategies",
          "universes",
          "incomplete_leg_ids",
          "n_legs_with_holdout",
          "n_universes_with_pairwise",
          "tables",
          "plots"
        ],
        "title": "StudyConsolidatedDTO",
        "type": "object"
      },
      "StudyDetail": {
        "properties": {
          "completed_legs": {
            "title": "Completed Legs",
            "type": "integer"
          },
          "completion_pct": {
            "title": "Completion Pct",
            "type": "number"
          },
          "cross_strategy_compares_done": {
            "items": {
              "type": "string"
            },
            "title": "Cross Strategy Compares Done",
            "type": "array"
          },
          "has_consolidated_report": {
            "title": "Has Consolidated Report",
            "type": "boolean"
          },
          "launched_by_username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Launched By Username"
          },
          "legs": {
            "items": {
              "$ref": "#/components/schemas/LegStateRow"
            },
            "title": "Legs",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "spec_hash": {
            "title": "Spec Hash",
            "type": "string"
          },
          "spec_name": {
            "title": "Spec Name",
            "type": "string"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "total_legs": {
            "title": "Total Legs",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "spec_name",
          "spec_hash",
          "started_at",
          "total_legs",
          "completed_legs",
          "completion_pct",
          "cross_strategy_compares_done",
          "legs",
          "has_consolidated_report"
        ],
        "title": "StudyDetail",
        "type": "object"
      },
...
      "StudySummary": {
        "properties": {
          "completed_legs": {
            "title": "Completed Legs",
            "type": "integer"
          },
          "completion_pct": {
            "title": "Completion Pct",
            "type": "number"
          },
          "launched_by_username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Launched By Username"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "spec_name": {
            "title": "Spec Name",
            "type": "string"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "total_legs": {
            "title": "Total Legs",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "spec_name",
          "started_at",
          "total_legs",
          "completed_legs",
          "completion_pct"
        ],
        "title": "StudySummary",
        "type": "object"
      },
      "TrialRow": {
        "properties": {
          "datetime_complete": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datetime Complete"
          },
          "datetime_start": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datetime Start"
          },
          "experiment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experiment Id"
          },
          "number": {
            "title": "Number",
            "type": "integer"
          },
          "params": {
            "additionalProperties": true,
            "title": "Params",
            "type": "object"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "number",
          "state",
          "value",
          "params",
          "datetime_start",
          "datetime_complete",
          "experiment_id"
        ],
        "title": "TrialRow",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Quant Webapp",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/hpo": {
      "get": {
        "operationId": "get_hpo_studies_api_hpo_get",
        "parameters": [
          {
            "in": "query",
            "name": "all",
            "required": false,
            "schema": {
              "default": false,
              "title": "All",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HpoSummary"
                  },
                  "title": "Response Get Hpo Studies Api Hpo Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Hpo Studies",
        "tags": [
          "hpo"
        ]
      }
    },
    "/api/hpo/{wire_id}": {
      "get": {
        "operationId": "get_hpo_study_detail_api_hpo__wire_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "wire_id",
            "required": true,
            "schema": {
              "title": "Wire Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HpoDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Hpo Study Detail",
        "tags": [
          "hpo"
        ]
      }
    },
    "/api/hpo/{wire_id}/param-importance": {
      "get": {
        "operationId": "get_hpo_param_importance_api_hpo__wire_id__param_importance_get",
        "parameters": [
          {
            "in": "path",
            "name": "wire_id",
            "required": true,
            "schema": {
              "title": "Wire Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParamImportanceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Hpo Param Importance",
        "tags": [
          "hpo"
        ]
      }
    },
    "/api/hpo/{wire_id}/trials": {
      "get": {
        "operationId": "get_hpo_trials_api_hpo__wire_id__trials_get",
        "parameters": [
          {
            "in": "path",
            "name": "wire_id",
            "required": true,
            "schema": {
              "title": "Wire Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after_trial",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After Trial"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TrialRow"
                  },
                  "title": "Response Get Hpo Trials Api Hpo  Wire Id  Trials Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Hpo Trials",
        "tags": [
          "hpo"
        ]
      }
    },
...
    "/api/studies": {
      "get": {
        "operationId": "get_studies_api_studies_get",
        "parameters": [
          {
            "in": "query",
            "name": "all",
            "required": false,
            "schema": {
              "default": false,
              "title": "All",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StudySummary"
                  },
                  "title": "Response Get Studies Api Studies Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Studies",
        "tags": [
          "studies"
        ]
      }
    },
    "/api/studies/{name}": {
      "get": {
        "operationId": "get_study_detail_api_studies__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Study Detail",
        "tags": [
          "studies"
        ]
      }
    },
    "/api/studies/{name}/consolidated": {
      "get": {
        "operationId": "get_study_consolidated_api_studies__name__consolidated_get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyConsolidatedDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Study Consolidated",
        "tags": [
          "studies"
        ]
      },
      "post": {
        "description": "Build (or rebuild) the consolidated report for a study and return it.\n\nRuns the matplotlib + table-writing work in the threadpool so the event\nloop stays responsive while the (few-seconds) job completes. A study\nthat hasn't completed any legs yet returns 422.",
        "operationId": "post_study_consolidated_api_studies__name__consolidated_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyConsolidatedDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Study Consolidated",
        "tags": [
          "studies"
        ]
      }
    },
    "/api/studies/{name}/consolidated/plots/{plot_name}": {
      "get": {
        "operationId": "get_study_consolidated_plot_api_studies__name__consolidated_plots__plot_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "plot_name",
            "required": true,
            "schema": {
              "title": "Plot Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Study Consolidated Plot",
        "tags": [
          "studies"
        ]
      }
    },
    "/api/studies/{name}/consolidated/tables/{table_name}": {
      "get": {
        "operationId": "get_study_consolidated_table_api_studies__name__consolidated_tables__table_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "table_name",
            "required": true,
            "schema": {
              "title": "Table Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Study Consolidated Table",
        "tags": [
          "studies"
        ]
      }
    },
...
  }
}